set up testing
