+++ simulate/samples.go
@@ -1,4 +1,4 @@
-package sampler
+package simulate
 
 // This script samples one number from each distribution
 // to construct the prevalence, positive row of the confusion matrix
